backend/app/config.py
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

class Settings:
    """Application settings loaded from environment variables"""
    
    # MongoDB Configuration
    MONGODB_URI: str = os.getenv("MONGODB_URI", "")
    DATABASE_NAME: str = os.getenv("DATABASE_NAME", "PersonalCFO")
    
    # API Configuration
    API_HOST: str = os.getenv("API_HOST", "0.0.0.0")
    API_PORT: int = int(os.getenv("API_PORT", "8000"))
    
    # Security
    SECRET_KEY: str = os.getenv("SECRET_KEY", "dev-secret-key-change-in-production")
    
    def validate(self):
        """Validate that required settings are present"""
        if not self.MONGODB_URI or self.MONGODB_URI == "":
            raise ValueError(
                "MONGODB_URI not found in environment variables. "
                "Please create a .env file in the backend/ directory with your MongoDB connection string."
            )
        return True

# Create a global settings instance
settings = Settings()
