backend/app/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-jose
email-validator